Move `Env` into `Utils`
