+++ src/main/scala/com/gradle/internal/CiUtils.scala
@@ -1,5 +1,7 @@
 package com.gradle.internal
 
+import com.gradle.internal.Utils.Env
+
 object CiUtils {
 
   private[gradle] def isCi(implicit env: Env): Boolean =
